@@ -1,34 +1,9 @@
 {
-  "_from": "@nodelib/fs.stat@^1.1.2",
-  "_id": "@nodelib/fs.stat@1.1.3",
-  "_inBundle": false,
-  "_integrity": "sha512-shAmDyaQC4H92APFoIaVDHCx5bStIocgvbwQyxPRrbUY20V1EYTbSDchWbuwlMG3V17cprZhA6+78JfB+3DTPw==",
-  "_location": "/@nodelib/fs.stat",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "@nodelib/fs.stat@^1.1.2",
-    "name": "@nodelib/fs.stat",
-    "escapedName": "@nodelib%2ffs.stat",
-    "scope": "@nodelib",
-    "rawSpec": "^1.1.2",
-    "saveSpec": null,
-    "fetchSpec": "^1.1.2"
-  },
-  "_requiredBy": [
-    "/fast-glob"
-  ],
-  "_resolved": "https://registry.npmjs.org/@nodelib/fs.stat/-/fs.stat-1.1.3.tgz",
-  "_shasum": "2b5a3ab3f918cca48a8c754c08168e3f03eba61b",
-  "_spec": "@nodelib/fs.stat@^1.1.2",
-  "_where": "/Users/harikiranyalavarthi/Documents/MultiGetApp/node_modules/fast-glob",
-  "bundleDependencies": false,
-  "deprecated": false,
+  "name": "@nodelib/fs.stat",
+  "version": "2.0.5",
   "description": "Get the status of a file with some features",
-  "engines": {
-    "node": ">= 6"
-  },
+  "license": "MIT",
+  "repository": "https://github.com/nodelib/nodelib/tree/master/packages/fs/fs.stat",
   "keywords": [
     "NodeLib",
     "fs",
@@ -36,22 +11,27 @@
     "file system",
     "stat"
   ],
-  "license": "MIT",
-  "main": "out/index.js",
-  "name": "@nodelib/fs.stat",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/nodelib/nodelib/tree/master/packages/fs/fs.stat"
+  "engines": {
+    "node": ">= 8"
   },
+  "files": [
+    "out/**",
+    "!out/**/*.map",
+    "!out/**/*.spec.*"
+  ],
+  "main": "out/index.js",
+  "typings": "out/index.d.ts",
   "scripts": {
-    "build": "npm run clean && npm run lint && npm run compile && npm test",
-    "clean": "rimraf out",
+    "clean": "rimraf {tsconfig.tsbuildinfo,out}",
+    "lint": "eslint \"src/**/*.ts\" --cache",
     "compile": "tsc -b .",
     "compile:watch": "tsc -p . --watch --sourceMap",
-    "lint": "tslint \"src/**/*.ts\" -p . -t stylish",
     "test": "mocha \"out/**/*.spec.js\" -s 0",
+    "build": "npm run clean && npm run compile && npm run lint && npm test",
     "watch": "npm run clean && npm run compile:watch"
   },
-  "typings": "out/index.d.ts",
-  "version": "1.1.3"
+  "devDependencies": {
+    "@nodelib/fs.macchiato": "1.0.4"
+  },
+  "gitHead": "d6a7960d5281d3dd5f8e2efba49bb552d090f562"
 }
