@@ -1,72 +1,38 @@
 {
-  "_from": "array-union@^1.0.1",
-  "_id": "array-union@1.0.2",
-  "_inBundle": false,
-  "_integrity": "sha1-mjRBDk9OPaI96jdb5b5w8kd47Dk=",
-  "_location": "/array-union",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "array-union@^1.0.1",
-    "name": "array-union",
-    "escapedName": "array-union",
-    "rawSpec": "^1.0.1",
-    "saveSpec": null,
-    "fetchSpec": "^1.0.1"
-  },
-  "_requiredBy": [
-    "/globby"
-  ],
-  "_resolved": "https://registry.npmjs.org/array-union/-/array-union-1.0.2.tgz",
-  "_shasum": "9a34410e4f4e3da23dea375be5be70f24778ec39",
-  "_spec": "array-union@^1.0.1",
-  "_where": "/Users/harikiranyalavarthi/Documents/MultiGetApp/node_modules/globby",
-  "author": {
-    "name": "Sindre Sorhus",
-    "email": "sindresorhus@gmail.com",
-    "url": "sindresorhus.com"
-  },
-  "bugs": {
-    "url": "https://github.com/sindresorhus/array-union/issues"
-  },
-  "bundleDependencies": false,
-  "dependencies": {
-    "array-uniq": "^1.0.1"
-  },
-  "deprecated": false,
-  "description": "Create an array of unique values, in order, from the input arrays",
-  "devDependencies": {
-    "ava": "*",
-    "xo": "*"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  },
-  "files": [
-    "index.js"
-  ],
-  "homepage": "https://github.com/sindresorhus/array-union#readme",
-  "keywords": [
-    "array",
-    "arr",
-    "set",
-    "uniq",
-    "unique",
-    "duplicate",
-    "remove",
-    "union",
-    "combine",
-    "merge"
-  ],
-  "license": "MIT",
-  "name": "array-union",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/sindresorhus/array-union.git"
-  },
-  "scripts": {
-    "test": "xo && ava"
-  },
-  "version": "1.0.2"
+	"name": "array-union",
+	"version": "2.1.0",
+	"description": "Create an array of unique values, in order, from the input arrays",
+	"license": "MIT",
+	"repository": "sindresorhus/array-union",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "sindresorhus.com"
+	},
+	"engines": {
+		"node": ">=8"
+	},
+	"scripts": {
+		"test": "xo && ava && tsd"
+	},
+	"files": [
+		"index.js",
+		"index.d.ts"
+	],
+	"keywords": [
+		"array",
+		"set",
+		"uniq",
+		"unique",
+		"duplicate",
+		"remove",
+		"union",
+		"combine",
+		"merge"
+	],
+	"devDependencies": {
+		"ava": "^1.4.1",
+		"tsd": "^0.7.2",
+		"xo": "^0.24.0"
+	}
 }
