@@ -89,23 +61,146 @@
     "promise",
     "node"
   ],
+  "author": "Matt Zabriskie",
   "license": "MIT",
-  "main": "index.js",
-  "name": "axios",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/axios/axios.git"
+  "bugs": {
+    "url": "https://github.com/axios/axios/issues"
   },
-  "scripts": {
-    "build": "NODE_ENV=production grunt build",
-    "coveralls": "cat coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
-    "examples": "node ./examples/server.js",
-    "postversion": "git push && git push --tags",
-    "preversion": "npm test",
-    "start": "node ./sandbox/server.js",
-    "test": "grunt test && bundlesize",
-    "version": "npm run build && grunt version && git add -A dist && git add CHANGELOG.md bower.json package.json"
+  "homepage": "https://axios-http.com",
+  "devDependencies": {
+    "@babel/core": "^7.18.2",
+    "@babel/preset-env": "^7.18.2",
+    "@commitlint/cli": "^17.3.0",
+    "@commitlint/config-conventional": "^17.3.0",
+    "@release-it/conventional-changelog": "^5.1.1",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-commonjs": "^15.1.0",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-multi-entry": "^4.0.0",
+    "@rollup/plugin-node-resolve": "^9.0.0",
+    "abortcontroller-polyfill": "^1.7.3",
+    "auto-changelog": "^2.4.0",
+    "body-parser": "^1.20.0",
+    "chalk": "^5.2.0",
+    "coveralls": "^3.1.1",
+    "cross-env": "^7.0.3",
+    "dev-null": "^0.1.1",
+    "dtslint": "^4.2.1",
+    "es6-promise": "^4.2.8",
+    "eslint": "^8.17.0",
+    "express": "^4.18.1",
+    "formdata-node": "^5.0.0",
+    "formidable": "^2.0.1",
+    "fs-extra": "^10.1.0",
+    "get-stream": "^3.0.0",
+    "gulp": "^4.0.2",
+    "gzip-size": "^7.0.0",
+    "handlebars": "^4.7.7",
+    "husky": "^8.0.2",
+    "istanbul-instrumenter-loader": "^3.0.1",
+    "jasmine-core": "^2.4.1",
+    "karma": "^6.3.17",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-firefox-launcher": "^2.1.2",
+    "karma-jasmine": "^1.1.1",
+    "karma-jasmine-ajax": "^0.1.13",
+    "karma-rollup-preprocessor": "^7.0.8",
+    "karma-safari-launcher": "^1.0.0",
+    "karma-sauce-launcher": "^4.3.6",
+    "karma-sinon": "^1.0.5",
+    "karma-sourcemap-loader": "^0.3.8",
+    "minimist": "^1.2.7",
+    "mocha": "^10.0.0",
+    "multer": "^1.4.4",
+    "pretty-bytes": "^6.0.0",
+    "release-it": "^15.5.1",
+    "rollup": "^2.67.0",
+    "rollup-plugin-auto-external": "^2.0.0",
+    "rollup-plugin-bundle-size": "^1.0.3",
+    "rollup-plugin-terser": "^7.0.2",
+    "sinon": "^4.5.0",
+    "stream-throttle": "^0.1.3",
+    "string-replace-async": "^3.0.2",
+    "terser-webpack-plugin": "^4.2.3",
+    "typescript": "^4.8.4",
+    "url-search-params": "^0.10.0"
   },
+  "browser": {
+    "./lib/adapters/http.js": "./lib/helpers/null.js",
+    "./lib/platform/node/index.js": "./lib/platform/browser/index.js",
+    "./lib/platform/node/classes/FormData.js": "./lib/helpers/null.js"
+  },
+  "jsdelivr": "dist/axios.min.js",
+  "unpkg": "dist/axios.min.js",
   "typings": "./index.d.ts",
-  "version": "0.18.0"
+  "dependencies": {
+    "follow-redirects": "^1.15.0",
+    "form-data": "^4.0.0",
+    "proxy-from-env": "^1.1.0"
+  },
+  "bundlesize": [
+    {
+      "path": "./dist/axios.min.js",
+      "threshold": "5kB"
+    }
+  ],
+  "contributors": [
+    "Matt Zabriskie (https://github.com/mzabriskie)",
+    "Nick Uraltsev (https://github.com/nickuraltsev)",
+    "Jay (https://github.com/jasonsaayman)",
+    "Dmitriy Mozgovoy (https://github.com/DigitalBrainJS)",
+    "Emily Morehouse (https://github.com/emilyemorehouse)",
+    "Rubén Norte (https://github.com/rubennorte)",
+    "Justin Beckwith (https://github.com/JustinBeckwith)",
+    "Martti Laine (https://github.com/codeclown)",
+    "Xianming Zhong (https://github.com/chinesedfan)",
+    "Rikki Gibson (https://github.com/RikkiGibson)",
+    "Remco Haszing (https://github.com/remcohaszing)",
+    "Yasu Flores (https://github.com/yasuf)",
+    "Ben Carp (https://github.com/carpben)"
+  ],
+  "sideEffects": false,
+  "release-it": {
+    "git": {
+      "commitMessage": "chore(release): v${version}",
+      "push": true,
+      "commit": true,
+      "tag": true,
+      "requireCommits": false,
+      "requireCleanWorkingDir": false
+    },
+    "github": {
+      "release": true,
+      "draft": true
+    },
+    "npm": {
+      "publish": false,
+      "ignoreVersion": false
+    },
+    "plugins": {
+      "@release-it/conventional-changelog": {
+        "preset": "angular",
+        "infile": "CHANGELOG.md",
+        "header": "# Changelog"
+      }
+    },
+    "hooks": {
+      "before:init": "npm test",
+      "after:bump": "gulp version --bump ${version} && npm run build && npm run test:build:version && git add ./dist && git add ./package-lock.json",
+      "before:release": "npm run release:changelog:fix",
+      "after:release": "echo Successfully released ${name} v${version} to ${repo.repository}."
+    }
+  },
+  "commitlint": {
+    "rules": {
+      "header-max-length": [
+        2,
+        "always",
+        130
+      ]
+    },
+    "extends": [
+      "@commitlint/config-conventional"
+    ]
+  }
 }
