@@ -1,100 +1,42 @@
 {
-  "_from": "braces@^2.3.1",
-  "_id": "braces@2.3.2",
-  "_inBundle": false,
-  "_integrity": "sha512-aNdbnj9P8PjdXU4ybaWLK2IF3jc/EoDYbC7AazW6to3TRsfXxscC9UXOB5iDiEQrkyIbWp2SLQda4+QAa7nc3w==",
-  "_location": "/braces",
-  "_phantomChildren": {
-    "is-extendable": "0.1.1"
-  },
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "braces@^2.3.1",
-    "name": "braces",
-    "escapedName": "braces",
-    "rawSpec": "^2.3.1",
-    "saveSpec": null,
-    "fetchSpec": "^2.3.1"
-  },
-  "_requiredBy": [
-    "/micromatch"
+  "name": "braces",
+  "description": "Bash-like brace expansion, implemented in JavaScript. Safer than other brace expansion libs, with complete support for the Bash 4.3 braces specification, without sacrificing speed.",
+  "version": "3.0.2",
+  "homepage": "https://github.com/micromatch/braces",
+  "author": "Jon Schlinkert (https://github.com/jonschlinkert)",
+  "contributors": [
+    "Brian Woodward (https://twitter.com/doowb)",
+    "Elan Shanker (https://github.com/es128)",
+    "Eugene Sharygin (https://github.com/eush77)",
+    "hemanth.hm (http://h3manth.com)",
+    "Jon Schlinkert (http://twitter.com/jonschlinkert)"
   ],
-  "_resolved": "https://registry.npmjs.org/braces/-/braces-2.3.2.tgz",
-  "_shasum": "5979fd3f14cd531565e5fa2df1abfff1dfaee729",
-  "_spec": "braces@^2.3.1",
-  "_where": "/Users/harikiranyalavarthi/Documents/MultiGetApp/node_modules/micromatch",
-  "author": {
-    "name": "Jon Schlinkert",
-    "url": "https://github.com/jonschlinkert"
-  },
+  "repository": "micromatch/braces",
   "bugs": {
     "url": "https://github.com/micromatch/braces/issues"
   },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Brian Woodward",
-      "url": "https://twitter.com/doowb"
-    },
-    {
-      "name": "Elan Shanker",
-      "url": "https://github.com/es128"
-    },
-    {
-      "name": "Eugene Sharygin",
-      "url": "https://github.com/eush77"
-    },
-    {
-      "name": "hemanth.hm",
-      "url": "http://h3manth.com"
-    },
-    {
-      "name": "Jon Schlinkert",
-      "url": "http://twitter.com/jonschlinkert"
-    }
-  ],
-  "dependencies": {
-    "arr-flatten": "^1.1.0",
-    "array-unique": "^0.3.2",
-    "extend-shallow": "^2.0.1",
-    "fill-range": "^4.0.0",
-    "isobject": "^3.0.1",
-    "repeat-element": "^1.1.2",
-    "snapdragon": "^0.8.1",
-    "snapdragon-node": "^2.0.1",
-    "split-string": "^3.0.2",
-    "to-regex": "^3.0.1"
-  },
-  "deprecated": false,
-  "description": "Bash-like brace expansion, implemented in JavaScript. Safer than other brace expansion libs, with complete support for the Bash 4.3 braces specification, without sacrificing speed.",
-  "devDependencies": {
-    "ansi-cyan": "^0.1.1",
-    "benchmarked": "^2.0.0",
-    "brace-expansion": "^1.1.8",
-    "cross-spawn": "^5.1.0",
-    "gulp": "^3.9.1",
-    "gulp-eslint": "^4.0.0",
-    "gulp-format-md": "^1.0.0",
-    "gulp-istanbul": "^1.1.2",
-    "gulp-mocha": "^3.0.1",
-    "gulp-unused": "^0.2.1",
-    "is-windows": "^1.0.1",
-    "minimatch": "^3.0.4",
-    "mocha": "^3.2.0",
-    "noncharacters": "^1.1.0",
-    "text-table": "^0.2.0",
-    "time-diff": "^0.3.1",
-    "yargs-parser": "^8.0.0"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  },
+  "license": "MIT",
   "files": [
     "index.js",
     "lib"
   ],
-  "homepage": "https://github.com/micromatch/braces",
+  "main": "index.js",
+  "engines": {
+    "node": ">=8"
+  },
+  "scripts": {
+    "test": "mocha",
+    "benchmark": "node benchmark"
+  },
+  "dependencies": {
+    "fill-range": "^7.0.1"
+  },
+  "devDependencies": {
+    "ansi-colors": "^3.2.4",
+    "bash-path": "^2.0.1",
+    "gulp-format-md": "^2.0.0",
+    "mocha": "^6.1.1"
+  },
   "keywords": [
     "alpha",
     "alphabetical",
