@@ -1,71 +1,38 @@
 {
-  "_from": "dir-glob@2.0.0",
-  "_id": "dir-glob@2.0.0",
-  "_inBundle": false,
-  "_integrity": "sha512-37qirFDz8cA5fimp9feo43fSuRo2gHwaIn6dXL8Ber1dGwUosDrGZeCCXq57WnIqE4aQ+u3eQZzsk1yOzhdwag==",
-  "_location": "/dir-glob",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "dir-glob@2.0.0",
-    "name": "dir-glob",
-    "escapedName": "dir-glob",
-    "rawSpec": "2.0.0",
-    "saveSpec": null,
-    "fetchSpec": "2.0.0"
-  },
-  "_requiredBy": [
-    "/globby"
-  ],
-  "_resolved": "https://registry.npmjs.org/dir-glob/-/dir-glob-2.0.0.tgz",
-  "_shasum": "0b205d2b6aef98238ca286598a8204d29d0a0034",
-  "_spec": "dir-glob@2.0.0",
-  "_where": "/Users/harikiranyalavarthi/Documents/MultiGetApp/node_modules/globby",
-  "author": {
-    "name": "Kevin Mårtensson",
-    "email": "kevinmartensson@gmail.com",
-    "url": "github.com/kevva"
-  },
-  "bugs": {
-    "url": "https://github.com/kevva/dir-glob/issues"
-  },
-  "bundleDependencies": false,
-  "dependencies": {
-    "arrify": "^1.0.1",
-    "path-type": "^3.0.0"
-  },
-  "deprecated": false,
-  "description": "Convert directories to glob compatible strings",
-  "devDependencies": {
-    "ava": "*",
-    "del": "^3.0.0",
-    "make-dir": "^1.0.0",
-    "rimraf": "^2.5.0",
-    "xo": "*"
-  },
-  "engines": {
-    "node": ">=4"
-  },
-  "files": [
-    "index.js"
-  ],
-  "homepage": "https://github.com/kevva/dir-glob#readme",
-  "keywords": [
-    "convert",
-    "directory",
-    "extensions",
-    "files",
-    "glob"
-  ],
-  "license": "MIT",
-  "name": "dir-glob",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/kevva/dir-glob.git"
-  },
-  "scripts": {
-    "test": "xo && ava"
-  },
-  "version": "2.0.0"
+	"name": "dir-glob",
+	"version": "3.0.1",
+	"description": "Convert directories to glob compatible strings",
+	"license": "MIT",
+	"repository": "kevva/dir-glob",
+	"author": {
+		"name": "Kevin Mårtensson",
+		"email": "kevinmartensson@gmail.com",
+		"url": "github.com/kevva"
+	},
+	"engines": {
+		"node": ">=8"
+	},
+	"scripts": {
+		"test": "xo && ava"
+	},
+	"files": [
+		"index.js"
+	],
+	"keywords": [
+		"convert",
+		"directory",
+		"extensions",
+		"files",
+		"glob"
+	],
+	"dependencies": {
+		"path-type": "^4.0.0"
+	},
+	"devDependencies": {
+		"ava": "^2.1.0",
+		"del": "^4.1.1",
+		"make-dir": "^3.0.0",
+		"rimraf": "^2.5.0",
+		"xo": "^0.24.0"
+	}
 }
