@@ -1,116 +1,94 @@
 {
-  "_from": "fast-glob@^2.0.2",
-  "_id": "fast-glob@2.2.6",
-  "_inBundle": false,
-  "_integrity": "sha512-0BvMaZc1k9F+MeWWMe8pL6YltFzZYcJsYU7D4JyDA6PAczaXvxqQQ/z+mDF7/4Mw01DeUc+i3CTKajnkANkV4w==",
-  "_location": "/fast-glob",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "fast-glob@^2.0.2",
-    "name": "fast-glob",
-    "escapedName": "fast-glob",
-    "rawSpec": "^2.0.2",
-    "saveSpec": null,
-    "fetchSpec": "^2.0.2"
-  },
-  "_requiredBy": [
-    "/globby"
-  ],
-  "_resolved": "https://registry.npmjs.org/fast-glob/-/fast-glob-2.2.6.tgz",
-  "_shasum": "a5d5b697ec8deda468d85a74035290a025a95295",
-  "_spec": "fast-glob@^2.0.2",
-  "_where": "/Users/harikiranyalavarthi/Documents/MultiGetApp/node_modules/globby",
+  "name": "fast-glob",
+  "version": "3.2.12",
+  "description": "It's a very fast and efficient glob library for Node.js",
+  "license": "MIT",
+  "repository": "mrmlnc/fast-glob",
   "author": {
     "name": "Denis Malinochkin",
-    "url": "canonium.com"
-  },
-  "bugs": {
-    "url": "https://github.com/mrmlnc/fast-glob/issues"
-  },
-  "bundleDependencies": false,
-  "dependencies": {
-    "@mrmlnc/readdir-enhanced": "^2.2.1",
-    "@nodelib/fs.stat": "^1.1.2",
-    "glob-parent": "^3.1.0",
-    "is-glob": "^4.0.0",
-    "merge2": "^1.2.3",
-    "micromatch": "^3.1.10"
-  },
-  "deprecated": false,
-  "description": "Is a faster `node-glob` alternative",
-  "devDependencies": {
-    "@types/bash-glob": "^2.0.0",
-    "@types/compute-stdev": "^1.0.0",
-    "@types/easy-table": "^0.0.32",
-    "@types/execa": "^0.9.0",
-    "@types/glob": "^7.1.1",
-    "@types/glob-parent": "^3.1.0",
-    "@types/glob-stream": "^6.1.0",
-    "@types/globby": "^8.0.0",
-    "@types/is-glob": "^4.0.0",
-    "@types/merge2": "^1.1.4",
-    "@types/micromatch": "^3.1.0",
-    "@types/minimist": "^1.2.0",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.0",
-    "@types/rimraf": "^2.0.2",
-    "bash-glob": "^2.0.0",
-    "compute-stdev": "^1.0.0",
-    "easy-table": "^1.1.1",
-    "execa": "^0.9.0",
-    "fast-glob": "^2.2.0",
-    "glob": "^7.1.2",
-    "glob-stream": "^6.1.0",
-    "globby": "^8.0.1",
-    "minimist": "^1.2.0",
-    "mocha": "^5.2.0",
-    "rimraf": "^2.6.2",
-    "tiny-glob": "^0.2.3",
-    "tslint": "^5.11.0",
-    "tslint-config-mrmlnc": "^1.0.0",
-    "typescript": "^3.1.3"
+    "url": "https://mrmlnc.com"
   },
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=8.6.0"
   },
-  "homepage": "https://github.com/mrmlnc/fast-glob#readme",
+  "main": "out/index.js",
+  "typings": "out/index.d.ts",
+  "files": [
+    "out",
+    "!out/{benchmark,tests}",
+    "!out/**/*.map",
+    "!out/**/*.spec.*"
+  ],
   "keywords": [
     "glob",
     "patterns",
     "fast",
     "implementation"
   ],
-  "license": "MIT",
-  "main": "index.js",
-  "name": "fast-glob",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mrmlnc/fast-glob.git"
+  "devDependencies": {
+    "@nodelib/fs.macchiato": "^1.0.1",
+    "@types/compute-stdev": "^1.0.0",
+    "@types/easy-table": "^0.0.32",
+    "@types/glob": "^7.1.1",
+    "@types/glob-parent": "^5.1.0",
+    "@types/is-ci": "^2.0.0",
+    "@types/merge2": "^1.1.4",
+    "@types/micromatch": "^4.0.0",
+    "@types/minimist": "^1.2.0",
+    "@types/mocha": "^5.2.7",
+    "@types/node": "^12.7.8",
+    "@types/rimraf": "^2.0.2",
+    "@types/sinon": "^7.5.0",
+    "compute-stdev": "^1.0.0",
+    "easy-table": "^1.1.1",
+    "eslint": "^6.5.1",
+    "eslint-config-mrmlnc": "^1.1.0",
+    "execa": "^2.0.4",
+    "fast-glob": "^3.0.4",
+    "fdir": "^5.1.0",
+    "glob": "^7.1.4",
+    "is-ci": "^2.0.0",
+    "log-update": "^4.0.0",
+    "minimist": "^1.2.0",
+    "mocha": "^6.2.1",
+    "rimraf": "^3.0.0",
+    "sinon": "^7.5.0",
+    "tiny-glob": "^0.2.6",
+    "typescript": "^3.6.3"
+  },
+  "dependencies": {
+    "@nodelib/fs.stat": "^2.0.2",
+    "@nodelib/fs.walk": "^1.2.3",
+    "glob-parent": "^5.1.2",
+    "merge2": "^1.3.0",
+    "micromatch": "^4.0.4"
   },
   "scripts": {
-    "bench": "npm run build && npm run bench-async && npm run bench-sync",
-    "bench-async": "npm run bench-async-1 && npm run bench-async-5 && npm run bench-async-10 && npm run bench-async-50 && npm run bench-async-100",
-    "bench-async-1": "node ./out/benchmark --depth 1",
-    "bench-async-10": "node ./out/benchmark --depth 10",
-    "bench-async-100": "node ./out/benchmark --depth 100",
-    "bench-async-5": "node ./out/benchmark --depth 5",
-    "bench-async-50": "node ./out/benchmark --depth 50",
-    "bench-sync": "npm run bench-sync-1 && npm run bench-sync-5 && npm run bench-sync-10 && npm run bench-sync-50 && npm run bench-sync-100",
-    "bench-sync-1": "node ./out/benchmark --depth 1 --type sync",
-    "bench-sync-10": "node ./out/benchmark --depth 10 --type sync",
-    "bench-sync-100": "node ./out/benchmark --depth 100 --type sync",
-    "bench-sync-5": "node ./out/benchmark --depth 5 --type sync",
-    "bench-sync-50": "node ./out/benchmark --depth 50 --type sync",
-    "build": "npm run clean && npm run lint && npm run compile && npm test",
     "clean": "rimraf out",
+    "lint": "eslint \"src/**/*.ts\" --cache",
     "compile": "tsc",
-    "lint": "tslint \"src/**/*.ts\" -p . -t stylish",
-    "smoke": "mocha \"out/**/*.smoke.js\" -s 0",
     "test": "mocha \"out/**/*.spec.js\" -s 0",
-    "watch": "npm run clean && npm run lint & npm run compile -- --sourceMap --watch"
-  },
-  "typings": "index.d.ts",
-  "version": "2.2.6"
+    "smoke": "mocha \"out/**/*.smoke.js\" -s 0",
+    "smoke:sync": "mocha \"out/**/*.smoke.js\" -s 0 --grep \"\\(sync\\)\"",
+    "smoke:async": "mocha \"out/**/*.smoke.js\" -s 0 --grep \"\\(async\\)\"",
+    "smoke:stream": "mocha \"out/**/*.smoke.js\" -s 0 --grep \"\\(stream\\)\"",
+    "build": "npm run clean && npm run compile && npm run lint && npm test",
+    "watch": "npm run clean && npm run compile -- --sourceMap --watch",
+    "bench": "npm run bench-async && npm run bench-stream && npm run bench-sync",
+    "bench-async": "npm run bench-async-flatten && npm run bench-async-deep && npm run bench-async-partial-flatten && npm run bench-async-partial-deep",
+    "bench-stream": "npm run bench-stream-flatten && npm run bench-stream-deep && npm run bench-stream-partial-flatten && npm run bench-stream-partial-deep",
+    "bench-sync": "npm run bench-sync-flatten && npm run bench-sync-deep && npm run bench-sync-partial-flatten && npm run bench-sync-partial-deep",
+    "bench-async-flatten": "node ./out/benchmark --mode async --pattern \"*\"",
+    "bench-async-deep": "node ./out/benchmark --mode async --pattern \"**\"",
+    "bench-async-partial-flatten": "node ./out/benchmark --mode async --pattern \"{fixtures,out}/{first,second}/*\"",
+    "bench-async-partial-deep": "node ./out/benchmark --mode async --pattern \"{fixtures,out}/**\"",
+    "bench-stream-flatten": "node ./out/benchmark --mode stream --pattern \"*\"",
+    "bench-stream-deep": "node ./out/benchmark --mode stream --pattern \"**\"",
+    "bench-stream-partial-flatten": "node ./out/benchmark --mode stream --pattern \"{fixtures,out}/{first,second}/*\"",
+    "bench-stream-partial-deep": "node ./out/benchmark --mode stream --pattern \"{fixtures,out}/**\"",
+    "bench-sync-flatten": "node ./out/benchmark --mode sync --pattern \"*\"",
+    "bench-sync-deep": "node ./out/benchmark --mode sync --pattern \"**\"",
+    "bench-sync-partial-flatten": "node ./out/benchmark --mode sync --pattern \"{fixtures,out}/{first,second}/*\"",
+    "bench-sync-partial-deep": "node ./out/benchmark --mode sync --pattern \"{fixtures,out}/**\""
+  }
 }
