@@ -1,80 +1,38 @@
 {
-  "_from": "fill-range@^4.0.0",
-  "_id": "fill-range@4.0.0",
-  "_inBundle": false,
-  "_integrity": "sha1-1USBHUKPmOsGpj3EAtJAPDKMOPc=",
-  "_location": "/fill-range",
-  "_phantomChildren": {
-    "is-extendable": "0.1.1"
-  },
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "fill-range@^4.0.0",
-    "name": "fill-range",
-    "escapedName": "fill-range",
-    "rawSpec": "^4.0.0",
-    "saveSpec": null,
-    "fetchSpec": "^4.0.0"
-  },
-  "_requiredBy": [
-    "/braces"
+  "name": "fill-range",
+  "description": "Fill in a range of numbers or letters, optionally passing an increment or `step` to use, or create a regex-compatible range with `options.toRegex`",
+  "version": "7.0.1",
+  "homepage": "https://github.com/jonschlinkert/fill-range",
+  "author": "Jon Schlinkert (https://github.com/jonschlinkert)",
+  "contributors": [
+    "Edo Rivai (edo.rivai.nl)",
+    "Jon Schlinkert (http://twitter.com/jonschlinkert)",
+    "Paul Miller (paulmillr.com)",
+    "Rouven Weßling (www.rouvenwessling.de)",
+    "(https://github.com/wtgtybhertgeghgtwtg)"
   ],
-  "_resolved": "https://registry.npmjs.org/fill-range/-/fill-range-4.0.0.tgz",
-  "_shasum": "d544811d428f98eb06a63dc402d2403c328c38f7",
-  "_spec": "fill-range@^4.0.0",
-  "_where": "/Users/harikiranyalavarthi/Documents/MultiGetApp/node_modules/braces",
-  "author": {
-    "name": "Jon Schlinkert",
-    "url": "https://github.com/jonschlinkert"
-  },
+  "repository": "jonschlinkert/fill-range",
   "bugs": {
     "url": "https://github.com/jonschlinkert/fill-range/issues"
   },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "email": "wtgtybhertgeghgtwtg@gmail.com",
-      "url": "https://github.com/wtgtybhertgeghgtwtg"
-    },
-    {
-      "name": "Edo Rivai",
-      "email": "edo.rivai@gmail.com",
-      "url": "edo.rivai.nl"
-    },
-    {
-      "name": "Jon Schlinkert",
-      "email": "jon.schlinkert@sellside.com",
-      "url": "http://twitter.com/jonschlinkert"
-    },
-    {
-      "name": "Paul Miller",
-      "email": "paul+gh@paulmillr.com",
-      "url": "paulmillr.com"
-    }
-  ],
-  "dependencies": {
-    "extend-shallow": "^2.0.1",
-    "is-number": "^3.0.0",
-    "repeat-string": "^1.6.1",
-    "to-regex-range": "^2.1.0"
-  },
-  "deprecated": false,
-  "description": "Fill in a range of numbers or letters, optionally passing an increment or `step` to use, or create a regex-compatible range with `options.toRegex`",
-  "devDependencies": {
-    "ansi-cyan": "^0.1.1",
-    "benchmarked": "^1.0.0",
-    "gulp-format-md": "^0.1.12",
-    "minimist": "^1.2.0",
-    "mocha": "^3.2.0"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  },
+  "license": "MIT",
   "files": [
     "index.js"
   ],
-  "homepage": "https://github.com/jonschlinkert/fill-range",
+  "main": "index.js",
+  "engines": {
+    "node": ">=8"
+  },
+  "scripts": {
+    "test": "mocha"
+  },
+  "dependencies": {
+    "to-regex-range": "^5.0.1"
+  },
+  "devDependencies": {
+    "gulp-format-md": "^2.0.0",
+    "mocha": "^6.1.1"
+  },
   "keywords": [
     "alpha",
     "alphabetical",
@@ -95,26 +53,8 @@
     "regex",
     "sh"
   ],
-  "license": "MIT",
-  "main": "index.js",
-  "name": "fill-range",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jonschlinkert/fill-range.git"
-  },
-  "scripts": {
-    "test": "mocha"
-  },
   "verb": {
-    "related": {
-      "list": [
-        "braces",
-        "expand-range",
-        "micromatch",
-        "to-regex-range"
-      ]
-    },
-    "toc": true,
+    "toc": false,
     "layout": "default",
     "tasks": [
       "readme"
@@ -125,6 +65,5 @@
     "lint": {
       "reflinks": true
     }
-  },
-  "version": "4.0.0"
+  }
 }
