@@ -1,69 +1,28 @@
 {
-  "_from": "follow-redirects@^1.3.0",
-  "_id": "follow-redirects@1.6.1",
-  "_inBundle": false,
-  "_integrity": "sha512-t2JCjbzxQpWvbhts3l6SH1DKzSrx8a+SsaVf4h6bG4kOXUuPYS/kg2Lr4gQSb7eemaHqJkOThF1BGyjlUkO1GQ==",
-  "_location": "/follow-redirects",
-  "_phantomChildren": {
-    "ms": "2.0.0"
-  },
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "follow-redirects@^1.3.0",
-    "name": "follow-redirects",
-    "escapedName": "follow-redirects",
-    "rawSpec": "^1.3.0",
-    "saveSpec": null,
-    "fetchSpec": "^1.3.0"
-  },
-  "_requiredBy": [
-    "/axios"
-  ],
-  "_resolved": "https://registry.npmjs.org/follow-redirects/-/follow-redirects-1.6.1.tgz",
-  "_shasum": "514973c44b5757368bad8bddfe52f81f015c94cb",
-  "_spec": "follow-redirects@^1.3.0",
-  "_where": "/Users/harikiranyalavarthi/Documents/MultiGetApp/node_modules/axios",
-  "author": {
-    "name": "Ruben Verborgh",
-    "email": "ruben@verborgh.org",
-    "url": "https://ruben.verborgh.org/"
-  },
-  "bugs": {
-    "url": "https://github.com/follow-redirects/follow-redirects/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Olivier Lalonde",
-      "email": "olalonde@gmail.com",
-      "url": "http://www.syskall.com"
-    },
-    {
-      "name": "James Talmage",
-      "email": "james@talmage.io"
-    }
-  ],
-  "dependencies": {
-    "debug": "=3.1.0"
-  },
-  "deprecated": false,
+  "name": "follow-redirects",
+  "version": "1.15.2",
   "description": "HTTP and HTTPS modules that follow redirects.",
-  "devDependencies": {
-    "concat-stream": "^1.6.0",
-    "coveralls": "^3.0.2",
-    "eslint": "^4.19.1",
-    "express": "^4.16.2",
-    "mocha": "^5.0.0",
-    "nyc": "^11.8.0"
-  },
-  "engines": {
-    "node": ">=4.0"
-  },
+  "license": "MIT",
+  "main": "index.js",
   "files": [
     "*.js"
   ],
+  "engines": {
+    "node": ">=4.0"
+  },
+  "scripts": {
+    "test": "npm run lint && npm run mocha",
+    "lint": "eslint *.js test",
+    "mocha": "nyc mocha"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:follow-redirects/follow-redirects.git"
+  },
   "homepage": "https://github.com/follow-redirects/follow-redirects",
+  "bugs": {
+    "url": "https://github.com/follow-redirects/follow-redirects/issues"
+  },
   "keywords": [
     "http",
     "https",
@@ -73,23 +32,28 @@
     "location",
     "utility"
   ],
-  "license": "MIT",
-  "main": "index.js",
-  "name": "follow-redirects",
-  "nyc": {
-    "reporter": [
-      "lcov",
-      "text"
-    ]
+  "author": "Ruben Verborgh <ruben@verborgh.org> (https://ruben.verborgh.org/)",
+  "contributors": [
+    "Olivier Lalonde <olalonde@gmail.com> (http://www.syskall.com)",
+    "James Talmage <james@talmage.io>"
+  ],
+  "funding": [
+    {
+      "type": "individual",
+      "url": "https://github.com/sponsors/RubenVerborgh"
+    }
+  ],
+  "peerDependenciesMeta": {
+    "debug": {
+      "optional": true
+    }
   },
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/follow-redirects/follow-redirects.git"
-  },
-  "scripts": {
-    "lint": "eslint *.js test",
-    "mocha": "nyc mocha",
-    "test": "npm run lint && npm run mocha"
-  },
-  "version": "1.6.1"
+  "devDependencies": {
+    "concat-stream": "^2.0.0",
+    "eslint": "^5.16.0",
+    "express": "^4.16.4",
+    "lolex": "^3.1.0",
+    "mocha": "^6.0.2",
+    "nyc": "^14.1.1"
+  }
 }
